MAGECLOUD-955: Create Install/Update Deploy processes
