+++ src/Magento/MagentoCloud/Command/Deploy.php
@@ -111,11 +111,7 @@ class Deploy extends Command
 
         $this->env->log("Starting deploy.");
         $this->saveEnvironmentData();
-        if (!$this->isInstalled()) {
-            $this->installMagento();
-        } else {
-            $this->updateMagento();
-        }
+
         $this->staticContentDeploy();
         $this->disableGoogleAnalytics();
         $this->env->log("Deployment complete.");
@@ -206,90 +202,6 @@ class Deploy extends Command
     }
 
     /**
-     * Verifies is Magento installed based on install date in env.php
-     *
-     * @return bool
-     */
-    public function isInstalled()
-    {
-        $configFile = $this->getConfigFilePath();
-        $installed = false;
-
-        //1. from environment variables check if db exists and has tables
-        //2. check if core_config_data and setup_module tables exist
-        //3. check install date
-
-        $this->env->log('Checking if db exists and has tables');
-        $output = $this->executeDbQuery('SHOW TABLES');
-        if (is_array($output) && count($output) > 1) {
-            if (!in_array('core_config_data', $output) || !in_array('setup_module', $output)) {
-                $this->env->log('Missing either core_config_data or setup_module table');
-                exit(5);
-            } elseif (file_exists($configFile)) {
-                $data = include $configFile;
-                if (isset($data['install']) && isset($data['install']['date'])) {
-                    $this->env->log("Magento was installed on " . $data['install']['date']);
-                    $installed = true;
-                } else {
-                    $config['install']['date'] = date('r');
-                    $updatedConfig = '<?php' . "\n" . 'return ' . var_export($config, true) . ';';
-                    file_put_contents($configFile, $updatedConfig);
-                    $installed = true;
-                }
-            } else {
-                $this->env->execute('touch ' . Environment::MAGENTO_ROOT . $configFile);
-                $config['install']['date'] = date('r');
-                $updatedConfig = '<?php' . "\n" . 'return ' . var_export($config, true) . ';';
-                file_put_contents($configFile, $updatedConfig);
-                $installed = true;
-            }
-        }
-        return $installed;
-    }
-
-    /**
-     * Run Magento installation
-     */
-    public function installMagento()
-    {
-        $this->env->log("Installing Magento.");
-
-        $urlUnsecure = $this->urls['unsecure'][''];
-        $urlSecure = $this->urls['secure'][''];
-
-        $command =
-            "php ./bin/magento setup:install \
-            --session-save=db \
-            --cleanup-database \
-            --currency=$this->defaultCurrency \
-            --base-url=$urlUnsecure \
-            --base-url-secure=$urlSecure \
-            --language=$this->adminLocale \
-            --timezone=America/Los_Angeles \
-            --db-host=$this->dbHost \
-            --db-name=$this->dbName \
-            --db-user=$this->dbUser \
-            --backend-frontname=$this->adminUrl \
-            --admin-user=$this->adminUsername \
-            --admin-firstname=$this->adminFirstname \
-            --admin-lastname=$this->adminLastname \
-            --admin-email=$this->adminEmail \
-            --admin-password=$this->adminPassword";
-
-        if (strlen($this->dbPassword)) {
-            $command .= " \
-            --db-password=$this->dbPassword";
-        }
-
-        $command .= $this->verbosityLevel;
-
-        $this->env->execute($command);
-
-        $this->setSecureAdmin();
-        $this->updateConfig();
-    }
-
-    /**
      * Update secure admin
      */
     public function setSecureAdmin()
@@ -299,244 +211,6 @@ class Deploy extends Command
             "php ./bin/magento config:set web/secure/use_in_adminhtml 1";
         $command .= $this->verbosityLevel;
         $this->env->execute($command);
-    }
-
-    /**
-     * Update Magento configuration
-     */
-    private function updateMagento()
-    {
-        $this->env->log("File env.php contains installation date. Updating configuration.");
-        $this->updateConfig();
-        $this->setupUpgrade();
-        $this->clearCache();
-    }
-
-    private function updateConfig()
-    {
-        $this->env->log("Updating configuration from environment variables.");
-        $this->updateConfiguration();
-        $this->updateAdminCredentials();
-        $this->updateSolrConfiguration();
-        $this->updateUrls();
-    }
-
-    /**
-     * Update admin credentials
-     */
-    private function updateAdminCredentials()
-    {
-        $this->env->log("Updating admin credentials.");
-
-        // @codingStandardsIgnoreStart
-        $this->executeDbQuery("update admin_user set firstname = '$this->adminFirstname', lastname = '$this->adminLastname', email = '$this->adminEmail', username = '$this->adminUsername', password='{$this->generatePassword($this->adminPassword)}' where user_id = '1';");
-        // @codingStandardsIgnoreEnd
-    }
-
-    /**
-     * Update SOLR configuration
-     */
-    private function updateSolrConfiguration()
-    {
-        $this->env->log("Updating SOLR configuration.");
-
-        if ($this->solrHost !== null
-            && $this->solrPort !== null
-            && $this->solrPath !== null
-            && $this->solrHost !== null
-        ) {
-            // @codingStandardsIgnoreStart
-            $this->executeDbQuery("update core_config_data set value = '$this->solrHost' where path = 'catalog/search/solr_server_hostname' and scope_id = '0';");
-            $this->executeDbQuery("update core_config_data set value = '$this->solrPort' where path = 'catalog/search/solr_server_port' and scope_id = '0';");
-            $this->executeDbQuery("update core_config_data set value = '$this->solrScheme' where path = 'catalog/search/solr_server_username' and scope_id = '0';");
-            $this->executeDbQuery("update core_config_data set value = '$this->solrPath' where path = 'catalog/search/solr_server_path' and scope_id = '0';");
-            // @codingStandardsIgnoreEnd
-        }
-    }
-
-    /**
-     * Update secure and unsecure URLs
-     */
-    private function updateUrls()
-    {
-        if ($this->enableUpdateUrls) {
-            $this->env->log("Updating secure and unsecure URLs.");
-            foreach ($this->urls as $urlType => $urls) {
-                foreach ($urls as $route => $url) {
-                    $prefix = 'unsecure' === $urlType ? self::PREFIX_UNSECURE : self::PREFIX_SECURE;
-                    if (!strlen($route)) {
-                        // @codingStandardsIgnoreStart
-                        $this->executeDbQuery("update core_config_data set value = '$url' where path = 'web/$urlType/base_url' and scope_id = '0';");
-                        // @codingStandardsIgnoreEnd
-                        continue;
-                    }
-                    $likeKey = $prefix . $route . '%';
-                    $likeKeyParsed = $prefix . str_replace('.', '---', $route) . '%';
-                    // @codingStandardsIgnoreStart
-                    $this->executeDbQuery("update core_config_data set value = '$url' where path = 'web/$urlType/base_url' and (value like '$likeKey' or value like '$likeKeyParsed');");
-                    // @codingStandardsIgnoreEnd
-                }
-            }
-        } else {
-            $this->env->log("Skipping URL updates");
-        }
-    }
-
-
-    /**
-     * Run Magento setup upgrade
-     */
-    private function setupUpgrade()
-    {
-        $this->env->log("Saving disabled modules.");
-
-        if (file_exists(Environment::REGENERATE_FLAG)) {
-            $this->env->log("Removing .regenerate flag");
-            unlink(Environment::REGENERATE_FLAG);
-        }
-
-        try {
-            /* Enable maintenance mode */
-            $this->env->log("Enabling Maintenance mode.");
-            $this->env->execute("php ./bin/magento maintenance:enable {$this->verbosityLevel}");
-
-            $this->env->log("Running setup upgrade.");
-            $this->env->execute("php ./bin/magento setup:upgrade --keep-generated -n {$this->verbosityLevel}");
-
-            /* Disable maintenance mode */
-            $this->env->execute("php ./bin/magento maintenance:disable {$this->verbosityLevel}");
-            $this->env->log("Maintenance mode is disabled.");
-        } catch (\RuntimeException $e) {
-            $this->env->log($e->getMessage());
-            //Rollback required by database
-            exit(6);
-        }
-        if (file_exists(Environment::REGENERATE_FLAG)) {
-            $this->env->log("Removing .regenerate flag");
-            unlink(Environment::REGENERATE_FLAG);
-        }
-    }
-
-    /**
-     * Clear Magento file based cache
-     */
-    private function clearCache()
-    {
-        $this->env->log("Clearing application cache.");
-
-        $this->env->execute(
-            "php ./bin/magento cache:flush {$this->verbosityLevel}"
-        );
-    }
-
-    /**
-     * Update env.php file content
-     */
-    private function updateConfiguration()
-    {
-        $this->env->log("Updating env.php database configuration.");
-
-        $configFileName = $this->getConfigFilePath();
-
-        $config = include $configFileName;
-
-        $config['db']['connection']['default']['username'] = $this->dbUser;
-        $config['db']['connection']['default']['host'] = $this->dbHost;
-        $config['db']['connection']['default']['dbname'] = $this->dbName;
-        $config['db']['connection']['default']['password'] = $this->dbPassword;
-
-        $config['db']['connection']['indexer']['username'] = $this->dbUser;
-        $config['db']['connection']['indexer']['host'] = $this->dbHost;
-        $config['db']['connection']['indexer']['dbname'] = $this->dbName;
-        $config['db']['connection']['indexer']['password'] = $this->dbPassword;
-
-        if ($this->amqpHost !== null && $this->amqpPort !== null
-            && $this->amqpUser !== null && $this->amqpPasswd !== null) {
-            $config['queue']['amqp']['host'] = $this->amqpHost;
-            $config['queue']['amqp']['port'] = $this->amqpPort;
-            $config['queue']['amqp']['user'] = $this->amqpUser;
-            $config['queue']['amqp']['password'] = $this->amqpPasswd;
-            $config['queue']['amqp']['virtualhost'] = $this->amqpVirtualhost;
-            $config['queue']['amqp']['ssl'] = $this->amqpSsl;
-        } else {
-            $config = $this->removeAmqpConfig($config);
-        }
-
-        if ($this->redisHost !== null && $this->redisPort !== null) {
-            $this->env->log("Updating env.php Redis cache configuration.");
-            $config['cache'] = $this->getRedisCacheConfiguration();
-            $config['session'] = [
-                'save' => 'redis',
-                'redis' => [
-                    'host' => $this->redisHost,
-                    'port' => $this->redisPort,
-                    'database' => $this->redisSessionDb
-                ]
-            ];
-        } else {
-            $config = $this->removeRedisConfiguration($config);
-        }
-
-        $config['backend']['frontName'] = $this->adminUrl;
-
-        $config['resource']['default_setup']['connection'] = 'default';
-
-        $updatedConfig = '<?php' . "\n" . 'return ' . var_export($config, true) . ';';
-
-        file_put_contents($configFileName, $updatedConfig);
-    }
-
-    /**
-     * Remove AMQP configuration from env.php
-     *
-     * @param array $config
-     * @return array
-     */
-    private function removeAmqpConfig(array $config)
-    {
-        $this->env->log("Removing AMQP configuration from env.php.");
-        if (isset($config['queue']['amqp'])) {
-            if (count($config['queue']) > 1) {
-                unset($config['queue']['amqp']);
-            } else {
-                unset($config['queue']);
-            }
-        }
-
-        return $config;
-    }
-
-    /**
-     * Generates admin password using default Magento settings
-     */
-    private function generatePassword($password)
-    {
-        $saltLenght = 32;
-        $charsLowers = 'abcdefghijklmnopqrstuvwxyz';
-        $charsUppers = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ';
-        $charsDigits = '0123456789';
-        $randomStr = '';
-        $chars = $charsLowers . $charsUppers . $charsDigits;
-
-        // use openssl lib
-        for ($i = 0, $lc = strlen($chars) - 1; $i < $saltLenght; $i++) {
-            $bytes = openssl_random_pseudo_bytes(PHP_INT_SIZE);
-            $hex = bin2hex($bytes); // hex() doubles the length of the string
-            $rand = abs(hexdec($hex) % $lc); // random integer from 0 to $lc
-            $randomStr .= $chars[$rand]; // random character in $chars
-        }
-        $salt = $randomStr;
-        $version = 1;
-        $hash = hash('sha256', $salt . $password);
-
-        return implode(
-            ':',
-            [
-                $hash,
-                $salt,
-                $version
-            ]
-        );
     }
 
     /**
@@ -739,41 +413,4 @@ class Deploy extends Command
         ];
     }
 
-    /**
-     * Return full path to environment configuration file.
-     *
-     * @return string The path to configuration file
-     */
-    private function getConfigFilePath()
-    {
-        return Environment::MAGENTO_ROOT . 'app/etc/env.php';
-    }
-
-    /**
-     * Clears configuration from redis usages.
-     *
-     * @param array $config An array of application configuration
-     * @return array
-     */
-    private function removeRedisConfiguration($config)
-    {
-        $this->env->log("Removing redis cache and session configuration from env.php.");
-
-        if (isset($config['session']['save']) && $config['session']['save'] == 'redis') {
-            $config['session']['save'] = 'db';
-            if (isset($config['session']['redis'])) {
-                unset($config['session']['redis']);
-            }
-        }
-
-        if (isset($config['cache']['frontend'])) {
-            foreach ($config['cache']['frontend'] as $cacheName => $cacheData) {
-                if (isset($cacheData['backend']) && $cacheData['backend'] == 'Cm_Cache_Backend_Redis') {
-                    unset($config['cache']['frontend'][$cacheName]);
-                }
-            }
-        }
-
-        return $config;
-    }
 }
